% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestor_abbrvs.R
\name{ancestor_abbrvs}
\alias{ancestor_abbrvs}
\title{Return a string of ancestor abbreviations in the order of the ancestor vectors}
\usage{
ancestor_abbrvs(L)
}
\arguments{
\item{L}{desired length of the output}
}
\description{
Use this to put axis labels on plots, etc.
}
\examples{
ancestor_abbrvs(15)
}
\keyword{internal}
