% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestor_vectors.R
\name{ancestor_vectors}
\alias{ancestor_vectors}
\title{Return a vector of ancestor IDs for each individual}
\usage{
ancestor_vectors(indivs, ped, num_generations)
}
\arguments{
\item{indivs}{a vector of IDs of individuals whose ancestors you want listed}

\item{ped}{a pedigree that includes all the individuals in \code{indivs} and also
their ancestors. It should be a tibble with columns \code{kid}, \code{pa} \code{ma}.}

\item{num_generations}{how many generations do you wish to record for
each sampled individual.  1 means just the individual themselves (so, not very
interesting, and you likely wouldn't ever use it, but it turns out to be handy,
for the programming, to define that level as 1); 2 means up to
and including the parents; 3 means up to and including the grandparents; 4 means up to
and including the great grandparents; and so forth.}
}
\description{
The goal here ultimately is to reduce a large pedigree to only the individuals
that we need, by dint of them being fairly close in the pedigree to the a group
of individuals that were sampled.  For each member of indivs, this function returns
a vector of its ancestors going back \code{num_generations} generations.  The ordering
is simple: 1 is the indiv, 2 is pa, 3 is ma, then 4 is pa's pa, 5 is pa's ma, 6 is
ma's pa, and 7 is ma's ma, then 8 is pa's pa's pa, and so forth, back as many
generations as one needs to go.
}
