% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ancestor_vectors_cpp}
\alias{ancestor_vectors_cpp}
\title{Function to make a vector of all the ancestors of an individual out to n generations.}
\arguments{
\item{sv}{vector of sample indexes}

\item{nv}{vector of names of all samples}

\item{Ped}{pedigree struct}

\item{n}{the number of generations.  0 = self, 1 = parent, 2 = grandparent, etc.}
}
\description{
This is a replacement for the R implementation of
\code{ancestor_vectors()} which was too slow. This will get called
from with a C function in which the pedigree has been assembled.
}
\keyword{internal}
