% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_and_plot_ancestry_matrices.R
\name{basic_amm_plot}
\alias{basic_amm_plot}
\title{here is a function to plot a basic ancestry match matrix}
\usage{
basic_amm_plot(ATP, add_imps = FALSE, perimeter_width = 0.5)
}
\arguments{
\item{ATP}{a tibble that has ind_1, ind_2, and an ID column}
}
\description{
It is primarily for internal use
}
\keyword{internal}
