% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_and_plot_ancestry_matrices.R
\name{count_and_plot_ancestry_matrices}
\alias{count_and_plot_ancestry_matrices}
\title{Count up the number of different kinds of related pairs and make a plot}
\usage{
count_and_plot_ancestry_matrices(Pairs, nrow = 6, ncol = 5)
}
\arguments{
\item{Pairs}{a tibble like that returned by \code{compile_related_pairs()}.}

\item{nrow}{the number of rows to plot per page}

\item{ncol}{the number of columns to plot per page}
}
\description{
This counts up all the different types of pairwise relationships and
makes a plot (or multiple pages of them).
}
