% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_and_plot_mate_distribution.R
\name{count_and_plot_mate_distribution}
\alias{count_and_plot_mate_distribution}
\title{Count and plot the number of mates each individuals has produced offspring with}
\usage{
count_and_plot_mate_distribution(P)
}
\arguments{
\item{P}{the pedigree from the simulation, like that returned in the \code{pedigree} component
of the list returned by \code{slurp_spip()}.}
}
\description{
This just tallies up the information from the pedigree.  It will plot things
faceted by pop (over rows) and sexes (over columns).
}
