% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{empty_crp}
\alias{empty_crp}
\title{A tibble like that produced by \code{compile_related_pairs()}, but with zero rows}
\source{
Just created it from the example
}
\usage{
empty_crp
}
\description{
This is included here so that it can be used as the return value for
\code{compile_related_pairs()} when there are no relatives.
}
