% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_amms}
\alias{example_amms}
\title{a list of examples of ancestry-match matrices}
\source{
I just made these for some illustrative figures in the manuscript about this
R package.
}
\usage{
example_amms
}
\description{
This is a list of matrices with names that describe what they
represent. The names are Relationship_Number-of-Generations, like:
"Unrelated_2gen", "Self_2gen", "Unrelated_3gen", "Unrelated_4gen",
"Unrelated_1gen", "Father_Offspring_2gen", "FullSibling_2gen", "PatHalfSibling_2gen",
"FullCousin_2gen", "HalfAuntNiece_2gen".
}
