% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_add_zone_perimeters.R
\name{gg_add_zone_perimeters}
\alias{gg_add_zone_perimeters}
\title{Add perimeters around the relationship zones}
\usage{
gg_add_zone_perimeters(g, L, perisize = 0.5)
}
\arguments{
\item{g}{the original ggplot}

\item{L}{the number of rows (or columns) in the ancestor-match matrices}

\item{perisize}{the size of the line to color in the perimeters}
}
\description{
Pass it the original ggplot, and this will return it, with the perimeters added.
}
\keyword{internal}
