% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_census_by_year_age_sex.R
\name{ggplot_census_by_year_age_sex}
\alias{ggplot_census_by_year_age_sex}
\title{Just a simple plot function}
\usage{
ggplot_census_by_year_age_sex(census)
}
\arguments{
\item{census}{a tibble of census counts with columns \code{year} and
\code{age}, and then the counts of the different sexes in columns
named \code{male}, and \code{female}.}
}
\value{
a ggplot object
}
\description{
Easy to do, but quicker to have it wrapped up in a plot.
}
