% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_n0.005_results}
\alias{growth_n0.005_results}
\title{The result of running spip with growth -0.005 in the population-growth-with-spip vignette}
\source{
Simulation results
}
\usage{
growth_n0.005_results
}
\description{
This is stored as package data so that the vignette can be written
even if spip is not installed on the system.
}
