% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{half_first_cousin_amm}
\alias{half_first_cousin_amm}
\title{A half-first cousin ancestry match matrix}
\source{
Simply wrote this down.
}
\usage{
half_first_cousin_amm
}
\description{
Just a simple AMM to use in some examples
}
