% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_sib_kin_probs.R
\name{half_sib_kin_probs}
\alias{half_sib_kin_probs}
\title{compute the probability that a pair are half siblings}
\usage{
half_sib_kin_probs(y, a1, a2, N, surv_probs, prob_repro, rel_repro)
}
\arguments{
\item{y}{the difference in birth years}

\item{a1}{the age at sampling of one member of the pair}

\item{a2}{the age at sampling of another member of the pair}

\item{N}{total number of 2+ individuals in the population. This
can be a vector.}

\item{rel_repro}{The relative fecundities/reproductive potential of the individuals}

\item{prob_repo}{The age-specific (starting from 1) probability of reproducing in a year}
}
\description{
Just a quick function for some developmental work.  This is not meant
for prime time (and, in fact, most of the inference related stuff like this
is going to be in a differnet package).
Depends on population sizes and covariates. This is not really
vectorized, but if you pass it multiple N values, it will return
a vector, one for each N value. Note that we only expect this to
work well for individuals with different birth years.  When they
have the same birth year they could have been full siblings and
also variance in reproductive success comes more into play. This
assumes an equal sex ratio.
}
