% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int_pow2}
\alias{int_pow2}
\title{Exponentiation function for integer arguments}
\arguments{
\item{x}{the integer to raise 2 to.}
}
\description{
Solaris' compiler freaked hard on the pow function---couldn't figure
out whether it should return an int or a float or something, when
I used it to get very small powers of two. So I
am going to write a silly, simple function (that only gets used a couple
of times in an entire execution, and only with very
small arguments, so the fact that it is not super efficient
should not be a big problem).
}
\keyword{internal}
