% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leslie_from_spip.R
\name{leslie_from_spip}
\alias{leslie_from_spip}
\title{Return a Leslie-like matrix from the spip parameters}
\usage{
leslie_from_spip(P, C)
}
\arguments{
\item{P}{a named list of the spip parameters.}

\item{C}{the constant size of the newborn cohort each year}
}
\value{
This function returns a list with the following components:
\itemize{
\item \code{stable_age_distro_fem}: a vector of the expected number of females in each age
group of 0 up to MaxAge once the stable age distribution has been reached.  Note that
this corresponds to the PREKILL_CENSUS from spip.  If you remove the first element and
consider the stable age distribution of ages 1 to MaxAge, that would correspond to the
POSTKILL_CENSUS from spip.  Note that spip's values are random and those reported here
are expectations.
\item \code{stable_age_distro_male}: same as above, but for males.
}
}
\description{
Here we take the survival rates for females and males and the sex ratio,
as well as the annual new cohort size (assumed constant),
and we make a leslie-like matrix
to compute the stable age distribution.
}
