% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{necessary_generations}
\alias{necessary_generations}
\title{Minimum number of generations for the ancestry match matrix to include relationship types}
\source{
I simply defined these
}
\description{
This is an integer vector of the \code{num_generations} required in the ancestry vectors
in order to observe each type of relationship in the ancestry match matrix.  For example,
if num_generations < 1 you can't see \code{PO} or \code{Si}.  This is named according to the
relationship zone names.
}
