% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_grid_plot.R
\name{pairs_grid_plot}
\alias{pairs_grid_plot}
\title{plot all the related pairs in a facet grid}
\usage{
pairs_grid_plot(pairs, x_buffer = 0.05, y_buffer = 0.05)
}
\arguments{
\item{pairs}{the pairs tibble from slurp_spip()}

\item{x_buffer}{fraction of the unit interval on the left and right that
things don't get jittered into.}

\item{y_buffer}{fraction of the unit interval on the top and bottom that
things don't get jittered into.}
}
\value{
sends back a ggplot object
}
\description{
blah blah
}
