% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_offspring_pairs.R
\name{parent_offspring_pairs}
\alias{parent_offspring_pairs}
\title{Return parent-offspring pairs amongst the samples}
\usage{
parent_offspring_pairs(S, P)
}
\arguments{
\item{S}{the samples tibble with columns ID, sex, born_year, sampling_year}

\item{P}{the pedigree with columns year, kid, pa, ma}
}
\description{
Given the true pedigree and a list of sampled individuals
return a tibble of all the parent offspring pairs.  Note that
if the individuals
were sampled multiple times, each of those different pairs
will show up.  This allows the covariate information to be used
in the model as well.
}
\details{
In this case the parent is always going to be ID_1 and the offspring
will be ID_2, in the output.
The column pair_type in the output takes the value "PO" for parent-offspring.
}
