% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amm_from_matrix.R
\name{plot_amm_from_matrix}
\alias{plot_amm_from_matrix}
\title{plot an ancestry matrix (or multiple such matrices) from its (their) matrix form}
\usage{
plot_amm_from_matrix(X)
}
\arguments{
\item{X}{input tibble with a factor or character column \code{ID} that gives
the "name" of the ancestry matrix that will be used if you want to facet
over the values in \code{ID}. And also \code{X} must have a list column \code{anc_match_matrix} each
element of which is a logical ancestry match matrix.}
}
\description{
For illustration purposes, if you want to simply plot an ancestry
matrix (or several) to show particular values, then this is the
handy function for you.
}
