% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conn_comps.R
\name{plot_conn_comps}
\alias{plot_conn_comps}
\title{plot the graph showing the connected components}
\usage{
plot_conn_comps(Pairs)
}
\arguments{
\item{Pairs}{the tibble that comes out of \code{compile_related_pairs()}.
For this function it must have, at least the
columns \code{id_1}, \code{id_2}, and \code{dom_relat}.}
}
\description{
Simple a wrapper for some tidygraph/ggraph stuff
}
