% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_dfs.R
\name{prepare_for_dfs}
\alias{prepare_for_dfs}
\title{Prepare input for the DFS relative-finding algorithm}
\usage{
prepare_for_dfs(ped, samp_vec)
}
\arguments{
\item{ped}{the pedigree.  A tibble with columns \code{kid}, \code{pa}, and \code{ma}}

\item{samp_vec}{A character vector with the names of the individuals that were sampled}
}
\description{
Uses the pedigree to convert individual IDs into indexes and
compile information to make the data structures for the DFS
algorithm to find relatives.
}
\keyword{internal}
