% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{primary_ancestor_pairs}
\alias{primary_ancestor_pairs}
\title{Return a list of the indices of the primary shared ancestors}
\usage{
primary_ancestor_pairs(M)
}
\arguments{
\item{M}{an ancestry match matrix (it is a logical matrix)}
}
\value{
A list of pairs.  Each pair is the 1-based index of ancestor of ind_1, then
ind_2 of the primary shared ancestors.
}
\description{
This operates on an ancestry match matrix and uses a simple, divide-by_two
relationship between an ancestor and its descendants in the ordering
of an ancestry vector to determine which of the matching ancestors are secondary,
and then return the ones that are primary.
}
\examples{
# find primary ancestor pairs of example AMMs
lapply(example_amms, primary_ancestor_pairs)
}
\keyword{internal}
