% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ancestors_and_relatives}
\alias{rcpp_ancestors_and_relatives}
\title{function to test and use DFS stuff}
\usage{
rcpp_ancestors_and_relatives(L, n)
}
\arguments{
\item{L}{list of inputs}

\item{n}{the number of generations back to go when computing the ancestor vectors
and finding relatives.}
}
\description{
function to test and use DFS stuff
}
\keyword{internal}
