% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{relationship_zone_names}
\alias{relationship_zone_names}
\title{relationship zone names}
\source{
I simply defined these
}
\description{
A simple character vector of 15 relationship zones in the order they
are encountered when traversing an ancestry match matrix out to four
generations, ordered by contribution of a single ancestor in the zone
to kinship and, within that, increasing in the maximum number of meioses
between the shared ancestor and either of the two pair members.
}
