% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_zone_perimeters.R
\name{relationship_zone_perimeters}
\alias{relationship_zone_perimeters}
\title{Return the perimeters of all the relationship zones}
\usage{
relationship_zone_perimeters()
}
\description{
This is primarily for plotting a figure in the paper about this package,
showing where all the relationship zones are.  It merely cycles over the
possible relationships in \code{\link{relationship_zone_names}} and produces one or
two rows in a tibble for each that has the corners of the rectangle of that
zone in the columns xmin, xmax, ymin, and ymax.  It is designed to be overlaid
upon the ancestry_match_matrix plots. There are some additional columns give
us the midpoint of the area, etc.
}
