% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatives_from_ancestry_vectors.R
\name{relatives_from_ancestry_vectors}
\alias{relatives_from_ancestry_vectors}
\title{Find relatives amongst samples from their ancestor vectors}
\usage{
relatives_from_ancestry_vectors(A)
}
\arguments{
\item{A}{a tibble that has columns of \code{ID} and \code{ancestors}.}
}
\description{
Given the ancestry vectors of a collection of individuals
(like the set of all sampled individuals from a simulation)
this function just figures out who is a relative of whom.
Relatives are defined as individuals that share at least
one ID within their ancestry vectors.  The way this function
works is by, for each ID, making a vector of all the other IDs
that include that original ID in its ancestry vectors.  Then,
for each individuals, these vectors are catenated (one for each
anestor) then they are uniqued and sorted.
}
