% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relpair_conn_comps.R
\name{relpair_conn_comps}
\alias{relpair_conn_comps}
\title{Find connected components amongst the related pairs}
\usage{
relpair_conn_comps(Pairs)
}
\arguments{
\item{Pairs}{a tibble with the related pairs.  It must have columns id_1 and
id_2}
}
\description{
This is a little utility function that actually gets called from
within \code{compile_related_pairs()} so we leave it as an internal function.
This takes in the tibble Pairs, and then spits out a tibble that looks
just the same, except that after the id_1 and id_2 columns it now has a
column called \code{conn_comp} that gives the connected component index of the
two pair members.
}
\keyword{internal}
