% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spip.R
\name{run_spip}
\alias{run_spip}
\title{Run spip in a user-specified directory}
\usage{
run_spip(
  pars,
  dir = tempfile(),
  spip_seeds = ceiling(runif(2, 1, 1e+09)),
  num_pops = 1
)
}
\arguments{
\item{pars}{A named list of parameter values.}

\item{dir}{The directory to run it in.  Defaults to a temp directory,
which will be unique every time it is run.}

\item{spip_seeds}{a vector of two positive integers. These get written to the
file spip_seeds, which is used by spip to seed its random number generator.  By
default, R supplies these two integers from its own random number generator.  This
way reproducible results from spip can be obtained by calling \code{set.seed()} from within
R before calling \code{run_spip()}.  For the most part, the user should never really have
to directly supply a value for spip_seeds.}
}
\description{
This runs it in a directory and the output from stdout
goes into a big file spip_out.txt in that directory. Currently
this is pretty bare bones.
}
\details{
This creates a temporary directory and runs spip in that directory, redirecting
stdout and stderr to files.  It then processes the output using awk to create a collection
of files.  If spip throws an error, the contents of stderr are written to the screen to notify
the user of how to correct their input.
}
