% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{search_up}
\alias{search_up}
\title{Depth first search up the pedigree to N generations.}
\arguments{
\item{i}{the index of the node to call this on}

\item{c}{the current generation level.  0 = the first (i.e. the sampled individual)}

\item{n}{the number of generations back to go.  1 means go no further than the parents.
2 means go no further than the grandparents.}

\item{P}{the pedigree structure}

\item{C}{a reference to a character vector to which sampled individuals' IDs will get
pushed on.  At the end, we can unique them.}
}
\description{
Call this
with c = 0 for the original individual, and it will go back, n generations.
Unlike in R, this is 0-based.  So, n = 1 is parents, n = 2 is grandparents
and so on.
}
\keyword{internal}
