% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibling_pairs.R
\name{sibling_pairs}
\alias{sibling_pairs}
\title{Return the sibling pairs from amongst the samples}
\usage{
sibling_pairs(S, P)
}
\arguments{
\item{S}{the samples tibble with columns ID, sex, born_year, sampling_year}

\item{P}{the pedigree with columns year, kid, pa, ma}
}
\description{
Given the true pedigree and a list of sampled individuals
return a tibble of all the full_sibling and half-sibling pairs.  Note that
if the individuals
were sampled multiple times, each of those different pairs
will show up.  This allows the covariate information to be used
in the model as well.
}
\details{
The column pair_type shows FS for full siblings and MHS for maternal half-siblings
and PHS for paternal half siblings.  Those two categories cannot always
be resolved in practice (though with mtDNA sequences it might be possible...)
}
