% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurp_spip_OLD.R
\name{slurp_spip_OLD}
\alias{slurp_spip_OLD}
\title{Read in spip output and return as a list of tibbles}
\usage{
slurp_spip_OLD(sfile)
}
\arguments{
\item{sfile}{the path to the spip outfile}
}
\description{
Bare bones at the moment.  This returns the full pedigree
and the census sizes.  It also returns a full list of the
samples, and also a tibble that includes the self, full-sib,
half-sib, and parent-offspring pairs amongst the samples, as
determined from the pedigree.
}
