% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_1_slurped_results_100_loci}
\alias{species_1_slurped_results_100_loci}
\title{The result of running spip in the species_1_simulation vignette with 100 loci.}
\source{
Simulation results
}
\usage{
species_1_slurped_results_100_loci
}
\description{
This is stored as package data so that the vignette can be written
even if spip is not installed on the system. This particular version
stores the results of running \code{run_spip()} calling for 100 loci segregating
in the population, then slurping the results up with \code{slurp_spip()}.
}
