% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_1_slurped_results_1gen}
\alias{species_1_slurped_results_1gen}
\title{The result of running spip in the species_1_simulation vignette and slurping out with num_generations = 1.}
\source{
Simulation results
}
\usage{
species_1_slurped_results_1gen
}
\description{
This is stored as package data so that the vignette can be written
even if spip is not installed on the system.
}
