% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_2_life_history}
\alias{species_2_life_history}
\title{a list of life-history / life-table data for another hypothetical species}
\source{
Fill this in later
}
\description{
species_2 for examples.  Note that I just set the male and female
rates and parameters similar.  Will probably need to adjust this
once I start playing with it.
}
