% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spip_binary.R
\name{spip_binary}
\alias{spip_binary}
\title{file path to be used in a call to spip}
\usage{
spip_binary()
}
\description{
This version checks to make sure it is there and throws an
error with a suggestion of how to get it if it is not there.
}
\keyword{internal}
