% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spip_exists.R
\name{spip_exists}
\alias{spip_exists}
\title{return TRUE if spip exists where it should be installed.}
\usage{
spip_exists()
}
\value{
A single logical scalar, either TRUE or FALSE.
}
\description{
This just checks for the file where it gets installed with
\code{install_spip()}. This is exported so it can be used to control
the flow in the vignettes, so that vignettes can still be
written using saved results when spip is not installed (i.e.,
at CRAN, etc.)
}
\examples{
# will be FALSE unless spip has been externally installed
spip_exists()
}
