% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spip_help.R
\name{spip_help}
\alias{spip_help}
\title{print the abbreviated usage information from spip}
\usage{
spip_help()
}
\description{
This simply calls spip with the \code{--help} option.
}
\examples{
\dontrun{spip_help()}
}
