% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spip_help_full.R
\name{spip_help_full}
\alias{spip_help_full}
\title{print the full usage information from spip}
\usage{
spip_help_full()
}
\description{
This simply calls spip with the \code{--help-full} option.
}
\examples{
\dontrun{spip_help()}
}
