% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_survival_from_census.R
\name{summarize_survival_from_census}
\alias{summarize_survival_from_census}
\title{Summarize annual sex-and-age-specific survival rates from the census information}
\usage{
summarize_survival_from_census(
  census,
  fem_surv_probs = NULL,
  male_surv_probs = NULL,
  nbins = 10
)
}
\arguments{
\item{census}{a tibble of census counts with columns \code{year} and
\code{age}, and then the counts of the different sexes in columns
named \code{male}, and \code{female}.}

\item{fem_surv_probs}{a vector of the parameters used for the simulation.  If present
these are put on the histogram plots. If you provide one of these, you have to provide both.}

\item{male_surv_probs}{a vector of the parameters used for the simulation.  If present
these are put on the histogram plots.}

\item{nbins}{number of bins for the histograms}
}
\value{
A list with components:
\itemize{
\item \code{survival_tibble}  A tibble with all the survival fractions
\item \code{plot_histos_by_age_and_sex} A histogram of observed survival fractions by age for male and females
across all years of the simulation.
}
}
\description{
The prekill census in year t+1 is the post-kill census in year t, so
we can use the prekill census to record the realized fraction of individuals
of each age and sex that survived the death episode in each year.  In the
output survival in year t is the fraction of j-year olds in year t that
survive to be j+1 year-olds in year t+1.
}
