% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{three_pops_no_mig_slurped_results}
\alias{three_pops_no_mig_slurped_results}
\title{The result of running spip and slurping the output in the three population case with no migration}
\source{
Simulation results
}
\usage{
three_pops_no_mig_slurped_results
}
\description{
This is stored as package data so that the vignette can be written
even if spip is not installed on the system.
}
