% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm2tibble.R
\name{amm2tibble}
\alias{amm2tibble}
\title{convert an ancestry-matching matrix to a ggplot-able tibble}
\usage{
amm2tibble(M)
}
\arguments{
\item{M}{an ancestry-match matrix}
}
\value{
\code{amm2tibble()} returns a tibble with three columns:
\itemize{
\item \code{x}: the 1-based index of the row of input matrix,
\item \code{y}: the 1-based index of the column of the input matrix,
\item \code{amm}: the logical value (TRUE/FALSE) of the (x,y)-th cell of the input matrix.
}
}
\description{
This merely converts a matrix to a tibble that can be plotted
easily using ggplot.
}
\examples{
# convert one of the simple example AMMs to a tibble
amm2tibble(example_amms$Father_Offspring_2gen)
}
