% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm2tibble.R
\name{amm2tibble}
\alias{amm2tibble}
\title{convert an ancestry-matching matrix to a ggplot-able tibble}
\usage{
amm2tibble(M)
}
\arguments{
\item{M}{an ancestry-match matrix}
}
\description{
This merely converts a matrix to a tibble that can be plotted
easily using ggplot.
}
\examples{
# convert one of the simple example AMMs to a tibble
amm2tibble(example_amms$Father_Offspring_2gen)
}
