% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm2tibble.R
\name{amm2tibble}
\alias{amm2tibble}
\title{convert an ancestry-matching matrix to a ggplot-able tibble}
\usage{
amm2tibble(M)
}
\arguments{
\item{x}{an amm}
}
\description{
convert an ancestry-matching matrix to a ggplot-able tibble
}
