% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_census_by_year_age_sex.R
\name{ggplot_census_by_year_age_sex}
\alias{ggplot_census_by_year_age_sex}
\title{Just a simple plot function}
\usage{
ggplot_census_by_year_age_sex(census)
}
\arguments{
\item{census}{a tibble of census counts with columns \code{year} and
\code{age}, and then the counts of the different sexes in columns
named \code{male}, and \code{female}.}
}
\value{
\code{ggplot_census_by_year_age_sex()} returns a ggplot object which is a
stacked barplot with year on the x-axis,
counts on the y-axis with fill mapped to age.  It is facet-gridded
with sex in the columns and populations in the rows.
}
\description{
Easy to do, but quicker to have it wrapped up in a plot.
}
\examples{
# A single population example
g <- ggplot_census_by_year_age_sex(species_1_slurped_results$census_postkill)

# a three-population example
g3 <- ggplot_census_by_year_age_sex(three_pops_with_mig_slurped_results$census_postkill)
}
