% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recapture_pairs.R
\name{recapture_pairs}
\alias{recapture_pairs}
\title{Return the recaptures from amongst the samples}
\usage{
recapture_pairs(S)
}
\arguments{
\item{S}{the samples tibble with columns ID, sex, born_year, sampling_year}
}
\description{
Given the tibble of sampled individuals
return a tibble of all the recaptures.
This is a somewhat complicated operation
that expands recaptures to all possible pairs of recaptures. (i.e. if the
individual was caught on 3 different occasions there will be three separate pairs).
The column pair_type shows RC for recapture.
}
\details{
This is broken now that sampling in different life episodes is
demarcated as such. But I have left it in here unexported and internal
in case I want to pick it up at some point.
}
\keyword{internal}
