% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurp_spip.R
\name{slurp_spip}
\alias{slurp_spip}
\title{Read in the pedigree, census, and sampling information from the spip run}
\usage{
slurp_spip(dir, num_generations)
}
\arguments{
\item{dir}{the path to the directory where spip was run.  This is
returned by \code{run_spip()}.}

\item{num_generations}{how many generations back do you wish to consider
for find relatives of
each sampled individual.  0 means just the individual themselves (so, not very
interesting, and you likely wouldn't ever use it. 1 means up to
and including the parents; 2 means up to and including the grandparents; 3 means up to
and including the great grandparents; and so forth.}
}
\value{
A list of tibbles:
\itemize{
\item \code{pedigree}:
\item \code{census}:
\item \code{samples}:
}
}
\description{
This version assumes that the spip output has been processed into
three files: \code{spip_pedigree.tsv}, \code{spip_prekill_census.tsv}, and
\code{spip_samples.tsv}, inside the directory where \code{run_spip()} was run.
}
