% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_1_life_history}
\alias{species_1_life_history}
\title{a list of life-history / life-table data for a hypothetical species}
\source{
Just values that might be typical of a fish.
}
\usage{
species_1_life_history
}
\description{
species_1 for examples.
}
