% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_2_life_history}
\alias{species_2_life_history}
\title{a list of life-history / life-table data for another hypothetical species}
\source{
This is something used for simulation testing
}
\usage{
species_2_life_history
}
\description{
species_2 for examples.  Note that I just set the male and female
rates and parameters similar.
}
