% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spip_binary_path.R
\name{spip_binary_path}
\alias{spip_binary_path}
\title{return the path where spip should be in the R system paths}
\usage{
spip_binary_path()
}
\description{
It expects it to be in the R package directory after external installation.
}
\keyword{internal}
