% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spip_help.R
\name{spip_help}
\alias{spip_help}
\title{print the abbreviated usage information from spip}
\usage{
spip_help()
}
\value{
This returns the exit status of \code{spip} if spip is installed,
but the return value is of little use.  Mainly this is run for the side
effect of printing the \code{spip} help menu to the console.
}
\description{
This simply calls spip with the \code{--help} option.
}
\examples{
\dontrun{spip_help()}
}
