% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spip_help_full.R
\name{spip_help_full}
\alias{spip_help_full}
\title{print the full usage information from spip}
\usage{
spip_help_full()
}
\value{
This returns the exit status of \code{spip} if spip is installed,
but the return value is of little use.  Mainly this is run for the side
effect of printing the \code{spip} full help menu to the console.
}
\description{
This simply calls spip with the \code{--help-full} option.
}
\examples{
\dontrun{spip_help()}
}
